Move getObservables reqs into TheAlertsTable
