+++ frontend/src/components/Alerts/AlertTableExpansion.vue
@@ -4,7 +4,7 @@
 <template>
   <ul>
     <!-- List each observable with link to filter -->
-    <li v-for="obs of observables" :key="obs.value">
+    <li v-for="obs of props.observables" :key="obs.value">
       <span class="link-text" @click="filterByObservable(obs)">{{
         formatObservable(obs)
       }}</span>
@@ -15,36 +15,12 @@
 </template>
 
 <script setup>
-  import { ref, defineProps, onMounted } from "vue";
-
+  import { defineProps } from "vue";
   import NodeTagVue from "../Node/NodeTag.vue";
 
-  import { NodeTree } from "@/services/api/nodeTree";
-
-  const observables = ref(null);
-
-  onMounted(async () => {
-    await getObservables(props.uuid);
-  });
-
   const props = defineProps({
-    uuid: { type: String, required: true },
+    observables: { type: Array, required: true },
   });
-
-  const getObservables = async (uuid) => {
-    const unsortedObservables = await NodeTree.readNodesOfNodeTree(
-      [uuid],
-      "observable",
-    );
-
-    observables.value = unsortedObservables.sort((a, b) => {
-      if (a.type.value === b.type.value) {
-        return a.value < b.value ? -1 : 1;
-      } else {
-        return a.type.value < b.type.value ? -1 : 1;
-      }
-    });
-  };
 
   const formatObservable = (observable) => {
     return `${observable.type.value} : ${observable.value}`;
